refactor: cli_utils -> cli
